# db: optimize SeekGE/SeekLT for the noop case