@sumeerbhola 2021-02-19
@jbowens I used your suggestion, from a previous optimization PR, of replacing `lastPositioningOpIsSeekPrefixGE` with an enum.